Wait for Redis to start accepting connections
Previously the RedisInside.Redis ctor would return immediately after
invoking the Redis server executable. In the event that the server was
unable to start (e.g. bind failure on the given TCP port) the consumer
of RedisInside would not know about this failure until they attempted to
connect to it. Now RedisInside will throw if the Redis server does not
start up. This should make it easier for consumers to retry failed
operations.